+++ backend/src/services/usersService.js
@@ -2,7 +2,7 @@ const fs = require('fs').promises;
 const path = require('path');
 const bcrypt = require('bcrypt');
 
-const USERS_FILE = path.join(__dirname, '..', 'users.json');
+const USERS_FILE = path.join(__dirname, '..', 'data', 'users.json');
 
 let users = [];
